backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
lancedb==0.4.0
ollama==0.1.0
websockets==12.0
boto3
botocore
psutil
httpx